+++ src/Fable.Websockets.Suave/Server.fs
@@ -2,8 +2,7 @@ namespace Fable.Websockets
 
 module Server =    
     open Protocol
-    open Observables    
-    open System        
+    open System
 
     type OnConnectionEstablished<'serverProtocol, 'clientProtocol> = 
       CloseHandle -> IObservable<WebsocketEvent<'serverProtocol>> -> SendMessage<'clientProtocol> -> IDisposable